travis: don't install groonga-tokenizer-mecab package if it already exists
This is for GROONGA_MASTER=yes.

+++ tools/travis/install.sh
@@ -25,7 +25,10 @@ export GROONGA_NORMALIZER_MYSQL_MASTER=yes
 curl https://raw.github.com/groonga/groonga/master/data/travis/setup.sh | sh
 curl https://raw.github.com/groonga/groonga-normalizer-mysql/master/data/travis/setup.sh | sh
 curl https://raw.github.com/clear-code/cutter/master/data/travis/setup.sh | sh
-sudo apt-get -qq -y install groonga-tokenizer-mecab
+
+if [ ! -f /usr/lib/groonga/plugins/tokenizers/mecab.so ]; then
+    sudo apt-get -qq -y install groonga-tokenizer-mecab
+fi
 
 mkdir -p vendor
 cd vendor
